// cosmowcl/src/app/shared/models/api/cosmobusiness/rilascioLockRequest.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * cosmobusiness
 * Api per cosmobusiness
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Model con le informazioni per il rilascio di un lock
 */
export interface RilascioLockRequest { 
    codiceRisorsa: string;
    codiceOwner: string;
}